Reorganized ts architecture
