+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "roll4d6",
-	"version": "0.6.0",
+	"version": "0.6.1",
 	"private": true,
 	"dependencies": {
 		"@testing-library/jest-dom": "^5.16.4",
