+++ src/components/card-elements/npc-ability-score.tsx
