+++ src/consts/abilities.ts
@@ -1,0 +1,1 @@
+export const abilities: string[] = ['strength', 'dexterity', 'constitution', 'intelligence', 'wisdom', 'charisma'];
