+++ src/consts/dragonborn-colors.tsx
@@ -1,0 +1,1 @@
+export const dragonbornColors: string[] = ['black', 'blue', 'brass', 'bronze', 'copper', 'gold', 'green', 'red', 'silver', 'white'];
